+++ contracts/MorpherOracle.sol
@@ -36,7 +36,7 @@ contract MorpherOracle is Ownable {
     mapping(bytes32 => uint256) public goodFrom;
     mapping(bytes32 => uint256) public goodUntil;
 
-    mapping(bytes32 => bool) public ordercancellationRequested;
+    mapping(bytes32 => bool) public orderCancellationRequested;
 
 // ----------------------------------------------------------------------------------
 // Events
@@ -93,10 +93,11 @@ contract MorpherOracle is Ownable {
 
     event OrderCancelled(
         bytes32 indexed _orderId,
-        address indexed _sender
+        address indexed _sender,
+        address indexed _oracleAddress
         );
 
-    event OrdercancellationRequested(
+    event OrderCancellationRequestedEvent(
         bytes32 indexed _orderId,
         address indexed _sender
         );
@@ -322,22 +323,25 @@ contract MorpherOracle is Ownable {
     }
 
     function initiateCancelOrder(bytes32 _orderId) public {
-        require(ordercancellationRequested[_orderId] == false, "Aborting: Order was already canceled");
+        require(orderCancellationRequested[_orderId] == false, "Aborting: Order was already canceled.");
         (address userId, , , , , , ) = tradeEngine.getOrder(_orderId);
-        require(userId == msg.sender, "Aborting: Only the user can request an order cancellation");
-        ordercancellationRequested[_orderId] = true;
-        emit OrdercancellationRequested(_orderId, msg.sender);
+        require(userId == msg.sender, "Aborting: Only the user can request an order cancellation.");
+        orderCancellationRequested[_orderId] = true;
+        emit OrderCancellationRequestedEvent(_orderId, msg.sender);
 
     }
     // ----------------------------------------------------------------------------------
     // cancelOrder(bytes32  _orderId)
     // User or Administrator can cancel their own orders before the _callback has been executed
     // ----------------------------------------------------------------------------------
-    function cancelOrder(bytes32 _orderId) public {
-        tradeEngine.cancelOrder(_orderId, msg.sender);
+    function cancelOrder(bytes32 _orderId) public onlyOracleOperator {
+        require(orderCancellationRequested[_orderId] == true, "Aborting: Order-Cancellation was not requested.");
+        (address userId, , , , , , ) = tradeEngine.getOrder(_orderId);
+        tradeEngine.cancelOrder(_orderId, userId);
         clearOrderConditions(_orderId);
         emit OrderCancelled(
             _orderId,
+            userId,
             msg.sender
             );
     }
