+++ test/morpherTradeEngine.basicTests.test.js
